requirements.txt
absl-py==2.3.0
beautifulsoup4==4.13.4
blinker==1.9.0
bs4==0.0.2
captcha==0.7.1
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
filelock==3.18.0
Flask==3.1.1
fonttools==4.58.1
fsspec==2025.5.1
grpcio==1.72.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.8
Markdown==3.8
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
netron==8.3.6
networkx==3.5
numpy==2.2.6
onnx==1.18.0
packaging==25.0
pillow==11.2.1
protobuf==6.31.1
pyasn1==0.6.1
pyparsing==3.2.3
python-dateutil==2.9.0.post0
requests==2.32.3
rsa==4.9.1
six==1.17.0
soupsieve==2.7
sympy==1.14.0
tensorboard==2.19.0
tensorboard-data-server==0.7.2
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
typing_extensions==4.13.2
urllib3==2.4.0
waitress==3.0.2
Werkzeug==3.1.3
